+++ src/component/Grid/Grid.tsx
@@ -1,16 +1,11 @@
-import Card from '../Card/Card'
-import { ToastContainer } from 'react-toastify'
-import 'react-toastify/dist/ReactToastify.css'
-import style from '../Styles/Grid.module.css'
-import { useSelector } from 'react-redux'
-import { ProductInterface } from '../../Interface'
-import { RootStateType } from '../../store/rootReducer'
 
-export default function Grid() {
-  
-  const {cardDetailsArray} = useSelector((state: RootStateType) => state.card);
+import React, {  ReactNode } from 'react'
 
+import style from './Grid.module.css'
+ import cartPageStyle from './CartEmpty.module.css'
 
+function GridTemplate({ children }:{children:ReactNode[]}) {
+ 
   return (
     <>
       <div
@@ -18,22 +13,17 @@ export default function Grid() {
         id='GridBox'
         style={{ float: 'left' }}
       >
-        {cardDetailsArray?.map((item: ProductInterface, index: number) => {
-          return <Card key={index} product={{ ...item, index: index }} />
-        })}{' '}
+        {children?.length ? (
+          children
+        ) : (
+          <div id={cartPageStyle['CartPageOuterDiv']}>
+            <hr />
+            <h1 id={cartPageStyle['CartPageIsEmpty']}>List is Empty</h1>
+          </div>
+        )}
       </div>
-      <ToastContainer
-        position='bottom-center'
-        autoClose={500}
-        hideProgressBar={false}
-        newestOnTop={false}
-        closeOnClick
-        rtl={false}
-        pauseOnFocusLoss
-        draggable
-        pauseOnHover
-        theme='colored'
-      />
     </>
   )
 }
+
+export default GridTemplate
