--- src/component/Grid/GridTemp.tsx
@@ -1,29 +1,0 @@
-
-import React, {  ReactNode } from 'react'
-
-import style from '../Styles/Grid.module.css'
-import { cartPageStyle} from '../'
-
-function GridTemplate({ children }:{children:ReactNode[]}) {
- 
-  return (
-    <>
-      <div
-        className={style['grid-container']}
-        id='GridBox'
-        style={{ float: 'left' }}
-      >
-        {children?.length ? (
-          children
-        ) : (
-          <div id={cartPageStyle['CartPageOuterDiv']}>
-            <hr />
-            <h1 id={cartPageStyle['CartPageIsEmpty']}>List is Empty</h1>
-          </div>
-        )}
-      </div>
-    </>
-  )
-}
-
-export default GridTemplate
